server/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
sqlalchemy==2.0.34
pydantic==2.9.2
python-dotenv==1.0.1
google-generativeai==0.8.3
numpy==2.1.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
psycopg2-binary==2.9.9